Instantiate interface types, incl. w/ methods that return type variable.
Interface types are instantiated on-the-fly via java.lang.reflect.Proxy.
When a method is invoked on a resulting proxy object, the associated
InvocationHandler recursively instantiates the method's return type. If
the return type is a type variable, the handler looks through the
parameters for that type variable. If found, it attempts to instantiate
the concrete class of the corresponding argument passed when the method
was invoked.