Process type arguments, use format strings in custom exceptions.
When a generic interface type needs to be instantiated to satisfy a
constructor parameter, its type arguments (for declared and inherited
type parameters) are recorded so that invocations of methods on the
resulting proxy object return the correct type, preventing
ClassCastExceptions. If an argument matching a return type type variable
can't be found, then the type variable must have been declared by the
method itself, in which case a parameter matching the type variable is
searched for and the type of the corresponding argument is used.

CloningFailedException and InstantiationFailedException now have varargs
Object parameters, passing the message and any Object arguments to
String.format(String,Object...) and passing the resulting formatted
string to a RuntimeException constructor, optionally with a Throwable
cause.